Fix : 참조오류 스크립트 수정
