+++ Assets/Externals/LowPolyFoodLite/LowPolyFoodLite_2020.3.0_HDRP_v1.0.unitypackage.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9fd45603780ce0549b97e796e74e5e3d
+guid: e8d1d59b4af53fb4685894e0972082fc
 DefaultImporter:
   externalObjects: {}
   userData:
