+++ Assets/Externals/LowPolyFoodLite/LowPolyFoodLite_2020.3.0_URP_v1.0.unitypackage.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 31ccef05f5ef9ff4e89cc18c68f4c794
+guid: 3294b408d2ae84a4ab6cf9b99f3f1628
 DefaultImporter:
   externalObjects: {}
   userData:
