+++ Assets/Externals/LowPolyRPGWeapons_Lite/LowPolyRPGWeaponsLite_2020.3.0_HDRP_v1.0.unitypackage.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 31ccef05f5ef9ff4e89cc18c68f4c794
+guid: dc011e675dff36648b315d5e9b8b76ec
 DefaultImporter:
   externalObjects: {}
   userData:
