+++ Assets/Externals/LowPolyRPGWeapons_Lite/LowPolyRPGWeaponsLite_2020.3.0_URP_v1.0.unitypackage.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 31ccef05f5ef9ff4e89cc18c68f4c794
+guid: 2c299b1162740664ba55a2fd6c2a8aee
 DefaultImporter:
   externalObjects: {}
   userData:
