+++ Assets/Resources/Prefabs/NPC.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 31ccef05f5ef9ff4e89cc18c68f4c794
-DefaultImporter:
+guid: 9ad2c1dc0e4b00c44b5b894afc8877c1
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
