+++ Assets/Resources/Prefabs/Player.prefab
@@ -1648,6 +1648,7 @@ GameObject:
   - component: {fileID: 5153912450459036203}
   - component: {fileID: 8760883662878191165}
   - component: {fileID: 4758623259296514912}
+  - component: {fileID: 1402830035465038961}
   m_Layer: 0
   m_Name: Player
   m_TagString: Untagged
@@ -1934,6 +1935,18 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: a318eef4667e0284c97fa3f656940a55, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &1402830035465038961
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2198478821790187540}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d6040df8e331d34498f641565d4364c4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1 &2221690031547684587
