--- Assets/Scenes/MainScene_501.unity.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-<<<<<<<< HEAD:Assets/Scenes/MainScene_YJK.unity.meta
-guid: f46d765eda7ac094aa537c4afe56e68e
-========
-guid: c05fc3b9d4ad78847a90d48ce1ab1478
->>>>>>>> dev:Assets/Scenes/JHK_MainScene.unity.meta
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
