--- Assets/Scenes/MainScene_YJK.unity.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-<<<<<<<< HEAD:Assets/Scenes/MainScene_YJK.unity.meta
-guid: 41605b655f5fd4f4bb334008da5bd7fa
-========
-guid: c05fc3b9d4ad78847a90d48ce1ab1478
->>>>>>>> dev:Assets/Scenes/JHK_MainScene.unity.meta
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
