--- Assets/Script/UI/Scene/UI_Test.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 650c47ac444d99845845a87059e3e615
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
