--- Assets/Script/UI/SupItem/UI_BuildSlot.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f051dc8ed5d6a174eadafe09674e8468
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
