--- Assets/Script/UI/SupItem/UI_ItemSlot.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4ef0818e193b35948997c8f5c9fcc7f7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
